On click scribbling effect
got the scribbling effect/animation working. It is only visible while left click is held down. Tried using "thread.sleep" to show off the animation but couldn't find a functional method. scribble matches the color of selected but due to this it blends into the scene, a black back ground many help this. effect follows mouse goal is to have it snap to objects and scale in size